upgrading to ltp latest

+++ testcases/kernel/security/selinux-testsuite/tests/file/selinux_file.sh
@@ -17,44 +17,54 @@ setup()
 	export TST_COUNT=0
 	export TST_TOTAL=14
 
-	# Clean up from a previous run
-	rm -f $SELINUXTMPDIR/temp_file 2>&1
-	rm -f $SELINUXTMPDIR/temp_file2 2>&1
-	rm -f $SELINUXTMPDIR/temp_file3 2>&1
+	LTPBIN=${LTPBIN:-$LTPROOT/testcases/bin}
+	SELINUXTMPDIR=$(mktemp -d)
+	chcon -t test_file_t $SELINUXTMPDIR
 
-	#
-	# Create the temp files
-	#
-	dd if=/dev/zero of=$SELINUXTMPDIR/temp_file count=2 ibs=1024 2>&1 > /dev/null
-	dd if=/dev/zero of=$SELINUXTMPDIR/temp_file2 count=2 ibs=1024 2>&1 > /dev/null
-	dd if=/dev/zero of=$SELINUXTMPDIR/temp_file3 count=2 ibs=1024 2>&1 > /dev/null
-	chmod 775 $SELINUXTMPDIR/temp_file 2>&1 > /dev/null
-	chmod 775 $SELINUXTMPDIR/temp_file2 2>&1 > /dev/null
+	if SELINUXTMPDIR=$(mktemp -d); then
 
-	#
-	# Change the context for the file the good domain only has access to.
-	#
-	chcon -t fileop_file_t $SELINUXTMPDIR/temp_file 2>&1 > /dev/null
+		chcon -t test_file_t $SELINUXTMPDIR
 
-	#
-	# Change the context for the r/w file for the bad domain
-	#
-	chcon -t nofileop_rw_file_t $SELINUXTMPDIR/temp_file2 2>&1 > /dev/null
+		#
+		# Create the temp files
+		#
+		dd if=/dev/zero of=$SELINUXTMPDIR/temp_file count=2 ibs=1024 2>&1 > /dev/null
+		dd if=/dev/zero of=$SELINUXTMPDIR/temp_file2 count=2 ibs=1024 2>&1 > /dev/null
+		dd if=/dev/zero of=$SELINUXTMPDIR/temp_file3 count=2 ibs=1024 2>&1 > /dev/null
+		chmod 775 $SELINUXTMPDIR/temp_file 2>&1 > /dev/null
+		chmod 775 $SELINUXTMPDIR/temp_file2 2>&1 > /dev/null
 
-	#
-	# Change the context for the read-only access file for the bad domain
-	#
-	chcon -t nofileop_ra_file_t $SELINUXTMPDIR/temp_file3 2>&1 > /dev/null
+		#
+		# Change the context for the file the good domain only has access to.
+		#
+		chcon -t fileop_file_t $SELINUXTMPDIR/temp_file 2>&1 > /dev/null
 
+		#
+		# Change the context for the r/w file for the bad domain
+		#
+		chcon -t nofileop_rw_file_t $SELINUXTMPDIR/temp_file2 2>&1 > /dev/null
+
+		#
+		# Change the context for the read-only access file for the bad domain
+		#
+		chcon -t nofileop_ra_file_t $SELINUXTMPDIR/temp_file3 2>&1 > /dev/null
+
+		# 
+		# Change the context of the test executable
+		#
+		chcon -t fileop_exec_t $LTPBIN/selinux_wait_io 2>&1 > /dev/null
+
+		#
+		# Get the SID of the good file.
+		#
+		good_file_sid=`ls -Z $SELINUXTMPDIR/temp_file | awk '{print $4}'`
+
+	fi
 	# 
 	# Change the context of the test executable
 	#
-	chcon -t fileop_exec_t $LTPBIN/selinux_wait_io 2>&1 > /dev/null
+	chcon -t fileop_exec_t selinux_wait_io 2>&1 > /dev/null
 
-	#
-	# Get the SID of the good file.
-	#
-	good_file_sid="system_u:object_r:fileop_file_t"
 }
 
 test01()
@@ -72,9 +82,9 @@ test01()
 	RC=$?
 	if [ $RC -ne 0 ]
 	then
-		echo "$TCID   FAIL : file failed."
+		tst_resm TFAIL "file failed."
 	else
-		echo "$TCID   PASS : file passed."
+		tst_resm TPASS "file passed."
 	fi
 	return $RC
 }
@@ -89,9 +99,9 @@ test02()
 	RC=$?
 	if [ $RC -ne 0 ]
 	then
-		echo "$TCID   FAIL : file failed."
+		tst_resm TFAIL "file failed."
         else
-		echo "$TCID   PASS : file passed."
+		tst_resm TPASS "file passed."
 	fi
 	return $RC
 }
@@ -106,9 +116,9 @@ test03()
 	RC=$?
 	if [ $RC -ne 0 ]
 	then
-		echo "$TCID   FAIL : file failed."
+		tst_resm TFAIL "file failed."
 	else
-		echo "$TCID   PASS : file passed."
+		tst_resm TPASS "file passed."
 	fi
 	return $RC
 }
@@ -123,9 +133,9 @@ test04()
 	RC=$?
 	if [ $RC -ne 0 ]
 	then
-		echo "$TCID   FAIL : file failed."
+		tst_resm TFAIL "file failed."
 	else
-		echo "$TCID   PASS : file passed."
+		tst_resm TPASS "file passed."
 	fi
 	return $RC
 }
@@ -140,9 +150,9 @@ test05()
 	RC=$?
 	if [ $RC -ne 0 ]
 	then
-		echo "$TCID   FAIL : file failed."
+		tst_resm TFAIL "file failed."
 	else
-		echo "$TCID   PASS : file passed."
+		tst_resm TPASS "file passed."
 	fi
 	return $RC
 }
@@ -157,9 +167,9 @@ test06()
 	RC=$?
 	if [ $RC -ne 0 ]
 	then
-		echo "$TCID   FAIL : file failed."
+		tst_resm TFAIL "file failed."
 	else
-		echo "$TCID   PASS : file passed."
+		tst_resm TPASS "file passed."
 	fi
 	return $RC
 }
@@ -181,20 +191,19 @@ test07()
 
 	# Run testcase in $LTPROOT/testcases/bin directory
 	SAVEPWD=${PWD}
-	cd ${LTPBIN}
-	CURRENTDIR="."
+	cd ${0%/*}
 
-	runcon -t test_fileop_t -- $CURRENTDIR/selinux_sigiotask 2>&1
+	runcon -t test_fileop_t -- selinux_sigiotask 2>&1
 	RC=$?
 	if [ $RC -ne 0 ]
 	then
-		echo "$TCID   FAIL : file failed."
+		tst_resm TFAIL "file failed."
 	else
-		echo "$TCID   PASS : file passed."
+		tst_resm TPASS "file passed."
 	fi
 	
 	# return to $LTPROOT directory
-	cd ${PWD}
+	cd ${SAVEPWD}
 
 	return $RC
 }
@@ -214,10 +223,10 @@ test08()
 	RC=$?
 	if [ $RC -ne 0 ]
 	then
-		echo "$TCID   PASS : file passed."
+		tst_resm TPASS "file passed."
 		RC=0
 	else
-		echo "$TCID   FAIL : file failed."
+		tst_resm TFAIL "file failed."
 		RC=1
 	fi
 	return $RC
@@ -243,9 +252,9 @@ test09()
 	RC=$?
 	if [ $RC -ne 0 ]
 	then
-		echo "$TCID   FAIL : file failed."
+		tst_resm TFAIL "file failed."
 	else
-		echo "$TCID   PASS : file passed."
+		tst_resm TPASS "file passed."
 	fi
 	return $RC
 }
@@ -260,10 +269,10 @@ test10()
 	RC=$?
 	if [ $RC -ne 0 ]
 	then
-		echo "$TCID   PASS : file passed."
+		tst_resm TPASS "file passed."
 		RC=0
 	else
-		echo "$TCID   FAIL : file failed."
+		tst_resm TFAIL "file failed."
 		RC=1
 	fi
 	return $RC
@@ -280,10 +289,10 @@ test11()
 	RC=$?
 	if [ $RC -ne 0 ]
 	then
-		echo "$TCID   PASS : file passed."
+		tst_resm TPASS "file passed."
 		RC=0
 	else
-		echo "$TCID   FAIL : file failed."
+		tst_resm TFAIL "file failed."
 		RC=1
 	fi
 	return $RC
@@ -300,10 +309,10 @@ test12()
 	RC=$?
 	if [ $RC -ne 0 ]
 	then
-		echo "$TCID   PASS : file passed."
+		tst_resm TPASS "file passed."
 		RC=0
 	else
-		echo "$TCID   FAIL : file failed."
+		tst_resm TFAIL "file failed."
 		RC=1
 	fi
 	return $RC
@@ -325,9 +334,9 @@ test13()
 	RC=$?
 	if [ $RC -ne 0 ]
 	then
-		echo "$TCID   FAIL : file failed."
+		tst_resm TFAIL "file failed."
 	else
-		echo "$TCID   PASS : file passed."
+		tst_resm TPASS "file passed."
 	fi
 	return $RC
 }
@@ -350,10 +359,10 @@ test14()
 	RC=$?
 	if [ $RC -ne 0 ]
 	then
-		echo "$TCID   PASS : file passed."
+		tst_resm TPASS "file passed."
 		RC=0
 	else
-		echo "$TCID   FAIL : file failed."
+		tst_resm TFAIL "file failed."
 		RC=1
 	fi
 	return $RC
@@ -361,12 +370,7 @@ test14()
 
 cleanup()
 {
-	#
-	# Delete the temp files
-	#
-	rm -f $basedir/temp_file 2>&1
-	rm -f $basedir/temp_file2 2>&1
-	rm -f $basedir/temp_file3 2>&1
+	rm -rf $SELINUXTMPDIR
 }
 
 #
@@ -380,20 +384,21 @@ cleanup()
 RC=0    # Return value from setup, and test functions.
 EXIT_VAL=0
 
-setup  
-test01 || EXIT_VAL=$RC
-test02 || EXIT_VAL=$RC
-test03 || EXIT_VAL=$RC
-test04 || EXIT_VAL=$RC
-test05 || EXIT_VAL=$RC
-test06 || EXIT_VAL=$RC
-test07 || EXIT_VAL=$RC
-test08 || EXIT_VAL=$RC
-test09 || EXIT_VAL=$RC
-test10 || EXIT_VAL=$RC
-test11 || EXIT_VAL=$RC
-test12 || EXIT_VAL=$RC
-test13 || EXIT_VAL=$RC
-test14 || EXIT_VAL=$RC
-cleanup
+if setup ; then
+	test01 || EXIT_VAL=$RC
+	test02 || EXIT_VAL=$RC
+	test03 || EXIT_VAL=$RC
+	test04 || EXIT_VAL=$RC
+	test05 || EXIT_VAL=$RC
+	test06 || EXIT_VAL=$RC
+	test07 || EXIT_VAL=$RC
+	test08 || EXIT_VAL=$RC
+	test09 || EXIT_VAL=$RC
+	test10 || EXIT_VAL=$RC
+	test11 || EXIT_VAL=$RC
+	test12 || EXIT_VAL=$RC
+	test13 || EXIT_VAL=$RC
+	test14 || EXIT_VAL=$RC
+	cleanup
+fi
 exit $EXIT_VAL
